Split JaCoCo badge generation into separate action
It is quite tricky to be able to run a workflow job that also acts as a
status check for branch protection when the job writes to the
repository.  Consider a PR: every time the workflow runs, a new commit
will be added that updates the branch’s state and this would prompt
another execution of the workflow, resulting in an infinite loop.
GitHub prevents this by not triggering Actions that would be triggered
by commits that were created by other Actions (note that I am
simplifying a bit here).

To work around this, let us not generate JaCoCo coverage badges in the
course of our single JaCoCo action.  Instead, let that action serve only
the purpose of allowing or forbidding a merge and reporting the coverage
in a PR comment.

Another action, defined in jacoco-badges.yml is going to actually
generate the badges and committing them, but only once the PR has been
merged into the main branch.