Allow empty commands
As of now, the command handler returned an error when an empty string
("") was supplied as a command.  That is, if you run the CLI and simply
press the enter key, you get an error message regarding the unknown ''
command.

This commit amends this and the CLI now accepts empty lines as no-op
commands.  After pressing enter, you are simply prompted again.

Coincidentally, this allows you to have empty lines in test files like
this:

------8<--------
GT4500,1,0,1,0

FIRE,SINGLE
FIRE,SINGLE

FIRE,ALL
------8<--------